refactor: Move notification actions into actions package
- Move `DownloadNotificationActions` into `app.actions`
- Move `AvailableUpdateNotificationActions` into `app.actions`
- Update imports to use `app.actions` package across codebase
- Remove old `DownloadNotificationActions` file
- Remove old `AvailableUpdateNotificationActions` file
- Update `DownloadWorker` to use `app.actions` imports
- Update `DownloadAvailableUpdateWorker` to use `app.actions` imports
- Update `DownloadNotificationActionReceiver` to use `app.actions` imports
- Update `AvailableUpdateNotificationActionReceiver` to use `AvailableUpdateNotificationActionType` from `app.actions`
- Add new `DownloadNotificationActions` implementation under `app.actions`
- Add new `AvailableUpdateNotificationActions` implementation under `app.actions`
