+++ app/src/main/kotlin/org/strigate/ferrot/app/receiver/DownloadNotificationActionReceiver.kt
@@ -13,17 +13,17 @@ import kotlinx.coroutines.launch
 import org.strigate.ferrot.R
 import org.strigate.ferrot.app.Constants.Extras.EXTRA_DOWNLOAD_ID
 import org.strigate.ferrot.app.Constants.Extras.EXTRA_NOTIFICATION_ACTION
-import org.strigate.ferrot.app.DownloadNotificationActionType
-import org.strigate.ferrot.app.DownloadNotificationActionType.DELETE
-import org.strigate.ferrot.app.DownloadNotificationActionType.MARK_SEEN
-import org.strigate.ferrot.app.DownloadNotificationActionType.RETRY
-import org.strigate.ferrot.app.DownloadNotificationActionType.STOP
-import org.strigate.ferrot.app.DownloadNotificationActionType.UNDO_DELETE
 import org.strigate.ferrot.app.NotificationService
-import org.strigate.ferrot.app.activeDownloadNotificationTag
-import org.strigate.ferrot.app.buildDownloadNotificationAction
-import org.strigate.ferrot.app.downloadNotificationExtras
-import org.strigate.ferrot.app.downloadNotificationTag
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType.DELETE
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType.MARK_SEEN
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType.RETRY
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType.STOP
+import org.strigate.ferrot.app.actions.DownloadNotificationActionType.UNDO_DELETE
+import org.strigate.ferrot.app.actions.activeDownloadNotificationTag
+import org.strigate.ferrot.app.actions.buildDownloadNotificationAction
+import org.strigate.ferrot.app.actions.downloadNotificationExtras
+import org.strigate.ferrot.app.actions.downloadNotificationTag
 import org.strigate.ferrot.domain.model.Download
 import org.strigate.ferrot.domain.model.DownloadStatus
 import org.strigate.ferrot.domain.usecase.DownloadMetadataUseCase
